+++ src/services/git.ts
@@ -4,7 +4,7 @@ import isEmpty from 'lodash.isempty';
 import uniq from 'lodash.uniq';
 import ora from 'ora';
 import { BackportOptions } from '../options/options';
-import { BackportCommit } from '../types/Commit';
+import { Commit } from '../types/Commit';
 import { HandledError } from './HandledError';
 import { execAsCallback, exec } from './child-process-promisified';
 import { getRepoOwnerPath, getRepoPath } from './env';
@@ -101,10 +101,7 @@ export async function addRemote(options: BackportOptions, remoteName: string) {
   }
 }
 
-export async function cherrypick(
-  options: BackportOptions,
-  commit: BackportCommit
-) {
+export async function cherrypick(options: BackportOptions, commit: Commit) {
   await exec(
     `git fetch ${options.repoOwner} ${commit.sourceBranch}:${commit.sourceBranch} --force`,
     { cwd: getRepoPath(options) }
@@ -151,10 +148,7 @@ export async function cherrypick(
   }
 }
 
-export async function commitChanges(
-  commit: BackportCommit,
-  options: BackportOptions
-) {
+export async function commitChanges(commit: Commit, options: BackportOptions) {
   const noVerify = options.noVerify ? ` --no-verify` : '';
 
   try {
