+++ src/services/github/v4/fetchCommitsByAuthor.ts
@@ -1,7 +1,7 @@
 import isEmpty from 'lodash.isempty';
 import ora from 'ora';
 import { BackportOptions } from '../../../options/options';
-import { BackportCommit } from '../../../types/Commit';
+import { Commit } from '../../../types/Commit';
 import { HandledError } from '../../HandledError';
 import {
   getFormattedCommitMessage,
@@ -9,18 +9,18 @@ import {
 } from '../commitFormatters';
 import { apiRequestV4 } from './apiRequestV4';
 import { fetchAuthorId } from './fetchAuthorId';
-import { getTargetBranchesFromLabels } from './getTargetBranchesFromLabels';
 import {
   pullRequestFragment,
   pullRequestFragmentName,
   PullRequestNode,
   getExistingTargetPullRequests,
   getPullRequestLabels,
-} from './sourcePRAndTargetPRs';
+} from './getExistingTargetPullRequests';
+import { getTargetBranchesFromLabels } from './getTargetBranchesFromLabels';
 
 export async function fetchCommitsByAuthor(
   options: BackportOptions
-): Promise<BackportCommit[]> {
+): Promise<Commit[]> {
   const {
     accessToken,
