+++ src/ui/getTargetBranches.ts
@@ -3,13 +3,10 @@ import isEmpty from 'lodash.isempty';
 import { BackportOptions } from '../options/options';
 import { HandledError } from '../services/HandledError';
 import { promptForTargetBranches } from '../services/prompts';
-import { BackportCommit } from '../types/Commit';
+import { Commit } from '../types/Commit';
 import { filterNil } from '../utils/filterEmpty';
 
-export function getTargetBranches(
-  options: BackportOptions,
-  commits: BackportCommit[]
-) {
+export function getTargetBranches(options: BackportOptions, commits: Commit[]) {
   // target branches already specified (in contrast to letting the user choose from a list)
   if (!isEmpty(options.targetBranches)) {
     return options.targetBranches;
